+++ brec_macros/src/lib.rs
@@ -1,5 +1,4 @@
 #![doc = include_str!("../README.md")]
-
 #[cfg(test)]
 #[cfg(feature = "generate_macro_test")]
 mod tests;