+++ examples/wasmjs/src/blocks/block_f32.rs
@@ -1,0 +1,7 @@
+use brec::prelude::*;
+
+#[block]
+#[derive(serde::Deserialize, serde::Serialize)]
+pub struct BlockF32 {
+    field: f32,
+}