+++ examples/wasmjs/src/blocks/block_i128.rs
@@ -1,0 +1,7 @@
+use brec::prelude::*;
+
+#[block]
+#[derive(serde::Deserialize, serde::Serialize)]
+pub struct BlockI128 {
+    field: i128,
+}