+++ examples/wasmjs/src/blocks/block_i16.rs
@@ -1,0 +1,7 @@
+use brec::prelude::*;
+
+#[block]
+#[derive(serde::Deserialize, serde::Serialize)]
+pub struct BlockI16 {
+    field: i16,
+}