+++ examples/wasmjs/src/blocks/block_i8.rs
@@ -1,0 +1,7 @@
+use brec::prelude::*;
+
+#[block]
+#[derive(serde::Deserialize, serde::Serialize)]
+pub struct BlockI8 {
+    field: i8,
+}