+++ examples/wasmjs/src/blocks/block_u128.rs
@@ -1,0 +1,7 @@
+use brec::prelude::*;
+
+#[block]
+#[derive(serde::Deserialize, serde::Serialize)]
+pub struct BlockU128 {
+    field: u128,
+}