+++ examples/wasmjs/src/blocks/block_u16.rs
@@ -1,0 +1,7 @@
+use brec::prelude::*;
+
+#[block]
+#[derive(serde::Deserialize, serde::Serialize)]
+pub struct BlockU16 {
+    field: u16,
+}