+++ examples/wasmjs/src/blocks/block_u32.rs
@@ -1,0 +1,7 @@
+use brec::prelude::*;
+
+#[block]
+#[derive(serde::Deserialize, serde::Serialize)]
+pub struct BlockU32 {
+    field: u32,
+}