+++ examples/wasmjs/src/blocks/block_u64.rs
@@ -1,0 +1,7 @@
+use brec::prelude::*;
+
+#[block]
+#[derive(serde::Deserialize, serde::Serialize)]
+pub struct BlockU64 {
+    field: u64,
+}