+++ examples/wasmjs/src/blocks/block_u8.rs
@@ -1,0 +1,7 @@
+use brec::prelude::*;
+
+#[block]
+#[derive(serde::Deserialize, serde::Serialize)]
+pub struct BlockU8 {
+    field: u8,
+}