+++ examples/wasmjs/src/payloads/d.rs
@@ -1,0 +1,17 @@
+use brec::prelude::*;
+
+#[payload(bincode)]
+#[derive(serde::Deserialize, serde::Serialize, PartialEq, PartialOrd, Debug, Clone)]
+pub enum PayloadD {
+    U8(u8),
+    U16(u16),
+    U32(u32),
+    U64(u64),
+    U128(u128),
+    I8(i8),
+    I16(i16),
+    I32(i32),
+    I64(i64),
+    I128(i128),
+    String(String),
+}