+++ examples/wasmjs/src/payloads/mod.rs
@@ -1,0 +1,9 @@
+mod a;
+mod b;
+mod c;
+mod d;
+
+pub use a::*;
+pub use b::*;
+pub use c::*;
+pub use d::*;