+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.1.2 (27.03.2025)
+
+### Fixes
+- Consider full path of derives on blocks and payloads parsing
+
+### Changes
+- Update documentation
+- Add wasm example
+- Include examples build into CI pipeline
+
 ## 0.1.1 (24.03.2025)
 
 ### Changes